Bugfix and improve rake task.
There were 2 issues regardless this rake task.

First one is about using assets path. It didn't took into consideration
assets.prefix path, it was hard coded as "/assets". This resulted in
ckeditor assets not being found. Now we use `Ckeditor.base_path` (which
uses `assets.prefix` or `Ckeditor.asset_path` if it was set). Besides
that we ensure that we are fetching files from within `Rails.root` path.

Second issue is about the way this task was appended to `asset:precompile`.

1. We need a separate task from within assets:precompile to standalone task
2. `Rake::Task#enhance` should be used to append it
3. Add configuration option if we want to run this task by hand

The separate task is mainly because of possible "race conditions" when
used with i.e. asset_sync gem combination, which is appending it's own
task as well to `assets:precompile` to push assets to CDN. When both
gems are appending it's tasks, we can't control the order in which they
run, and we need to ensure that ckeditor task needs to run first. So while
having a separate rake task `ckeditor:nondigest` we can specify which
task should run when, i.e.: `rake assets:precompile ckeditor:nondigest assets:sync`